hand merging with step_alias

+++ mflowgen/bin/setup-buildkite.sh
@@ -355,7 +355,7 @@ echo ""
 #    % multivt -> ../../../mflowgen/adks/tsmc16/multivt/
 
 mflowgen_orig=$mflowgen
-ln -s $mflowgen_orig
+test -e mflowgen || ln -s $mflowgen_orig
 mflowgen=`pwd`/mflowgen
